// client-svelte/src/lib/generated/apiClient.ts
export enum TorrentDownloadAction {
	DownloadAll = 0,
	DownloadAvailableFiles = 1,
	DownloadManual = 2
}

export enum TorrentFinishedAction {
	None = 0,
	RemoveAllTorrents = 1,
	RemoveRealDebrid = 2,
	RemoveClient = 3
}

export enum TorrentHostDownloadAction {
	DownloadAll = 0,
	DownloadNone = 1
}